+++ src/gibman.nim
@@ -4,8 +4,9 @@
 
 import cligen, config, osproc, utils
 
-const Version = """
-gibman 1.0.1
+const Version =
+  """
+gibman 2.0.0
 Program written by Maxwell Jensen (c) 2025
 Licensed under European Union Public Licence 1.2.
 For more information, consult README.md"""
@@ -28,12 +29,13 @@ proc argParser(list_presets = false, preset = "", verbose = false, version = fal
   process.close()
 
 dispatch argParser,
-  cmdName="gibman",
-  doc="A DOOM WAD manager for nerds",
-  help={
-  "help": "Print this information",
-  "help-syntax": "Program interface details",
-  "list-presets": "List presets defined in configuration file",
-  "preset": "Specify preset to run DOOM with",
-  "verbose": "Run program in verbose mode",
-  "version": "Print version information"}
+  cmdName = "gibman",
+  doc = "A DOOM WAD manager for nerds",
+  help = {
+    "help": "Print this information",
+    "help-syntax": "Program interface details",
+    "list-presets": "List presets defined in configuration file",
+    "preset": "Specify preset to run DOOM with",
+    "verbose": "Run program in verbose mode",
+    "version": "Print version information",
+  }