+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] 2025-03-16
+
+### Changed
+
+Removed `wad` from configuration file, replaced with `search`, which is a list
+of paths that gibman uses to search for WADs instead. The search is
+case-insensitive.
+
 ## [1.0.0] 2025-03-08
 
 ### Added